Make basic tests more robust.
In the daily CI of the PackageDistro repository, curlInterface
occasionally fails its tests. I believe this is because of the "dumb"
check for the result not containing a certain status code, e.g. 404.

Now, here is a typical result:

    {
      "args": {},
      "headers": {
        "Accept": "*/*",
        "Host": "www.httpbin.org",
        "User-Agent": "curlInterface/GAP package",
        "X-Amzn-Trace-Id": "Root=1-62f4fc1e-00401ff3692cd43d1a2234a5"
      },
      "origin": "131.246.132.154",
      "url": "http://www.httpbin.org/get"
    }

Note the hex string associated to the "X-Amzn-Trace-Id" key... And note
how in my example it actually contains "401" as a substring. Clearly
it can easily also contain "404". I believe this is why the curlInterface
tests sometimes fail.

As a workaround, just search for "404 " instead of "404". It won't match
this, but when we see a genuine 404 error page, it will contain a string
like "404 Not Found" or so.